+++ packages/fast-usdc-contract/test/contract-setup.ts
@@ -99,6 +99,9 @@ const startContract = async (
     installation,
     { USDC: usdc.issuer },
     { usdcDenom: uusdcOnAgoric },
+    // @ts-expect-error XXX contract expecting CosmosChainInfo with bech32
+    // prefix but the Orchestration setup doesn't have it. The tests pass anyway
+    // so we elide this infidelity to production.
     commonPrivateArgs,
   );
