+++ packages/fast-usdc-contract/test/fast-usdc.contract.test.ts
@@ -49,6 +49,9 @@ test('initial baggage', async t => {
     installation,
     { USDC: usdc.issuer },
     { usdcDenom: uusdcOnAgoric },
+    // @ts-expect-error XXX contract expecting CosmosChainInfo with bech32
+    // prefix but the Orchestration setup doesn't have it. The tests pass anyway
+    // so we elide this infidelity to production.
     commonPrivateArgs,
   );
